# Conflicts:
#	ts/WoltLabSuite/Core/Controller/Moderation/AssignUser.ts
#	wcfsetup/install/files/acp/js/WCF.ACP.js
#	wcfsetup/install/files/js/WoltLabSuite/Core/Controller/Moderation/AssignUser.js
#	wcfsetup/install/files/js/WoltLabSuite/Core/Media/Manager/Select.js
#	wcfsetup/install/files/js/WoltLabSuite/Core/Ui/Feed/Dialog.js
#	wcfsetup/install/files/lib/acp/action/InstallPackageAction.class.php
#	wcfsetup/install/files/lib/acp/form/AbstractCategoryEditForm.class.php
#	wcfsetup/install/files/lib/acp/form/LanguageItemAddForm.class.php
#	wcfsetup/install/files/lib/acp/form/LoginForm.class.php
#	wcfsetup/install/files/lib/acp/page/VersionTrackerListPage.class.php
#	wcfsetup/install/files/lib/data/blacklist/entry/BlacklistEntryAction.class.php
#	wcfsetup/install/files/lib/data/blacklist/status/BlacklistStatus.class.php
#	wcfsetup/install/files/lib/data/category/CategoryAction.class.php
#	wcfsetup/install/files/lib/data/label/Label.class.php
#	wcfsetup/install/files/lib/data/like/LikeAction.class.php
#	wcfsetup/install/files/lib/data/notice/Notice.class.php
#	wcfsetup/install/files/lib/data/notice/NoticeAction.class.php
#	wcfsetup/install/files/lib/data/paid/subscription/user/PaidSubscriptionUserAction.class.php
#	wcfsetup/install/files/lib/data/user/group/UserGroupAction.class.php
#	wcfsetup/install/files/lib/form/AccountManagementForm.class.php
#	wcfsetup/install/files/lib/form/ContactForm.class.php
#	wcfsetup/install/files/lib/form/RegisterForm.class.php
#	wcfsetup/install/files/lib/http/middleware/JsonBody.class.php
#	wcfsetup/install/files/lib/page/AbstractArticlePage.class.php
#	wcfsetup/install/files/lib/page/ArticleFeedPage.class.php
#	wcfsetup/install/files/lib/system/SingletonFactory.class.php
#	wcfsetup/install/files/lib/system/background/job/ServiceWorkerDeliveryBackgroundJob.class.php
#	wcfsetup/install/files/lib/system/bbcode/AttachmentBBCode.class.php
#	wcfsetup/install/files/lib/system/bbcode/UserBBCode.class.php
#	wcfsetup/install/files/lib/system/box/AbstractDatabaseObjectListBoxController.class.php
#	wcfsetup/install/files/lib/system/cronjob/CronjobScheduler.class.php
#	wcfsetup/install/files/lib/system/email/mime/AbstractMultipartMimePart.class.php
#	wcfsetup/install/files/lib/system/email/mime/MimePartFacade.class.php
#	wcfsetup/install/files/lib/system/endpoint/controller/core/moderationQueues/MarkModerationQueueAsRead.class.php
#	wcfsetup/install/files/lib/system/endpoint/controller/core/notices/DisableNotice.class.php
#	wcfsetup/install/files/lib/system/endpoint/controller/core/notices/EnableNotice.class.php
#	wcfsetup/install/files/lib/system/endpoint/controller/core/versionTrackers/RevertVersion.class.php
#	wcfsetup/install/files/lib/system/file/processor/FileProcessor.class.php
#	wcfsetup/install/files/lib/system/gridView/admin/ExceptionLogGridView.class.php
#	wcfsetup/install/files/lib/system/gridView/admin/UserGroupGridView.class.php
#	wcfsetup/install/files/lib/system/html/input/node/HtmlInputNodeProcessor.class.php
#	wcfsetup/install/files/lib/system/html/input/node/HtmlInputNodeWoltlabMetacode.class.php
#	wcfsetup/install/files/lib/system/html/metacode/converter/UrlMetacodeConverter.class.php
#	wcfsetup/install/files/lib/system/html/output/node/HtmlOutputNodeImg.class.php
#	wcfsetup/install/files/lib/system/interaction/admin/ArticleInteractions.class.php
#	wcfsetup/install/files/lib/system/interaction/admin/AttachmentInteractions.class.php
#	wcfsetup/install/files/lib/system/interaction/admin/CaptchaQuestionInteractions.class.php
#	wcfsetup/install/files/lib/system/interaction/bulk/admin/AttachmentBulkInteractions.class.php
#	wcfsetup/install/files/lib/system/io/Tar.class.php
#	wcfsetup/install/files/lib/system/io/Zip.class.php
#	wcfsetup/install/files/lib/system/message/embedded/object/ArticleMessageEmbeddedObjectHandler.class.php
#	wcfsetup/install/files/lib/system/message/quote/MessageQuoteManager.class.php
#	wcfsetup/install/files/lib/system/message/unfurl/UnfurlResponse.class.php
#	wcfsetup/install/files/lib/system/option/user/LabeledUrlUserOptionOutput.class.php
#	wcfsetup/install/files/lib/system/page/handler/ArticlePageHandler.class.php
#	wcfsetup/install/files/lib/system/request/RequestHandler.class.php
#	wcfsetup/install/files/lib/system/search/acp/OptionACPSearchResultProvider.class.php
#	wcfsetup/install/files/lib/system/style/FontManager.class.php
#	wcfsetup/install/files/lib/system/template/TemplateScriptingCompiler.class.php
#	wcfsetup/install/files/lib/system/template/plugin/IconFunctionTemplatePlugin.class.php
#	wcfsetup/install/files/lib/system/template/plugin/IpSearchModifierTemplatePlugin.class.php
#	wcfsetup/install/files/lib/system/template/plugin/JsonModifierTemplatePlugin.class.php
#	wcfsetup/install/files/lib/system/template/plugin/JsphraseFunctionTemplatePlugin.class.php
#	wcfsetup/install/files/lib/system/template/plugin/JsphrasePrefilterTemplatePlugin.class.php
#	wcfsetup/install/files/lib/system/user/authentication/password/algorithm/TPhpass.class.php
#	wcfsetup/install/files/lib/system/worker/SendNewPasswordWorker.class.php
#	wcfsetup/install/files/lib/util/IpAddress.class.php

+++ wcfsetup/install/files/lib/system/interaction/admin/LabelGroupInteractions.class.php
@@ -26,6 +26,10 @@ final class LabelGroupInteractions extends AbstractInteractionProvider
 {
     public function __construct()
     {
+        if (!WCF::getSession()->hasPermission('admin.content.label.canManageLabel')) {
+            return;
+        }
+
         $this->addInteractions([
             new DeleteInteraction('core/labels/groups/%s'),
             new class(
