+++ wcfsetup/install/files/lib/system/interaction/admin/MenuInteractions.class.php
@@ -28,6 +28,10 @@ final class MenuInteractions extends AbstractInteractionProvider
 {
     public function __construct()
     {
+        if (!WCF::getSession()->hasPermission('admin.content.cms.canManageMenu')) {
+            return;
+        }
+
         $this->addInteractions([
             new DeleteInteraction(
                 "core/menus/%s",
